requirements.txt
aenum==3.1.15
decorator==5.1.1
diskcache==5.6.3
evadb==0.3.7
greenlet==3.0.0
lark==1.1.7
numpy==1.26.0
pandas==2.1.1
psutil==5.9.5
py==1.11.0
pydantic==1.10.13
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
retry==0.9.2
six==1.16.0
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
thefuzz==0.20.0
typing_extensions==4.8.0
tzdata==2023.3
fastai==1.0.61
torch
torchvision~=0.16.0


Pillow~=10.1.0